requirements.txt
aiomysql==0.2.0
anyio==4.4.0
async-sqlalchemy==1.0.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.98.0
GeoAlchemy2==0.15.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
numpy==2.0.0
packaging==24.1
pydantic==1.10.9
PyMySQL==1.1.1
python-dotenv==0.21.0
shapely==2.0.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
